techdocs-node: inline google storage mock
